new_implementation/requirements.txt
# New Implementation Dependencies

# Core framework
pydantic>=2.0.0
aiohttp>=3.8.0
aiosqlite>=0.19.0
apscheduler>=3.10.0
pyyaml>=6.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data processing
pandas>=2.0.0
openpyxl>=3.1.0  # For Excel file support
odfpy>=1.4.0     # For ODS file support

# Optional integrations
discord.py>=2.3.0    # For Discord notifications
python-telegram-bot>=20.0  # For Telegram notifications
playwright>=1.40.0   # For browser automation

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0